requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1
django-allauth==64.1.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-filter==24.3
environs==11.0.0
marshmallow==3.22.0
packaging==24.1
psycopg==3.1.19
psycopg-binary==3.1.19
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2